backendfastapi/requirements.txt
fastapi
uvicorn[standard]
requests
sqlmodel
psycopg[binary]
python-decouple